# apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: |
    OpenAPI specification for the marketplace backend.

    The API uses a consistent envelope format:
    - successful responses: `{ "data": ... }`
    - errors: `{ "error": { "code", "message", "details" } }`
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: /
    description: Relative server URL for reverse proxy setups
...
paths:
...
  /api/v1/auth/verify-email/request:
    post:
      tags: [Auth]
      operationId: requestEmailVerification
      summary: Request a new email verification link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
            examples:
              default:
                value:
                  email: user@example.com
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedEnvelope'
        '400':
          $ref: '#/components/responses/InvalidInputResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
  /api/v1/auth/verify-email/confirm:
    post:
      tags: [Auth]
      operationId: confirmEmailVerification
      summary: Confirm email verification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailConfirmRequest'
            examples:
              default:
                value:
                  token: verify_3e85ce8092e64c79a7e0b0b15da20148
      responses:
        '200':
          description: Email confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailEnvelope'
        '400':
          $ref: '#/components/responses/InvalidTokenResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
  /api/v1/auth/password-reset/request:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      summary: Request a password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
            examples:
              default:
                value:
                  email: user@example.com
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedEnvelope'
        '400':
          $ref: '#/components/responses/InvalidInputResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [Auth]
      operationId: confirmPasswordReset
      summary: Confirm password reset token and set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
            examples:
              default:
                value:
                  token: reset_89e2ec8252f24b9fa8f0fd22ad4e2f0d
                  new_password: NewSecure123
      responses:
        '200':
          description: Password reset completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetEnvelope'
        '400':
          $ref: '#/components/responses/InvalidTokenResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/v1/auth/logout-all:
    post:
      tags: [Auth]
      operationId: logoutAll
      summary: Revoke all refresh sessions for the current user
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      responses:
        '200':
          description: All refresh sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokedAllEnvelope'
        '403':
          $ref: '#/components/responses/CSRFInvalidResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/v1/auth/sessions/{id}:
    delete:
      tags: [Auth]
      operationId: revokeSession
      summary: Revoke a specific refresh session for the current user
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIDPathParam'
        - $ref: '#/components/parameters/CSRFHeader'
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokedEnvelope'
        '400':
          $ref: '#/components/responses/InvalidInputResponse'
        '403':
          $ref: '#/components/responses/CSRFInvalidResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: mp_access_token
  headers:
    RetryAfterHeader:
      description: Number of seconds until the client may retry the request.
      schema:
        type: integer
        minimum: 1
        example: 60
  examples:
    RateLimitedExample:
      value:
        error:
          code: rate_limited
          message: too many requests
          details:
            scope: auth_login
            retry_after_seconds: 60
...
  parameters:
...
    CSRFHeader:
      name: X-CSRF-Token
      in: header
      description: Required for mutating requests when cookie auth mode is enabled.
      required: false
      schema:
        type: string
        minLength: 20
      example: csrf_1f9f7b7c92a14c8fb6debdfe94eebfef
...
    SessionIDPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 4d019535-6f30-4dc9-a96f-0647d14318c1
  responses:
    InvalidInputResponse:
      description: Invalid request payload or malformed parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalidInput:
              value:
                error:
                  code: invalid_input
                  message: invalid input
    RateLimitedResponse:
      description: Too many requests for the same endpoint from the same client.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfterHeader'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            rateLimited:
              $ref: '#/components/examples/RateLimitedExample'
    InvalidTokenResponse:
      description: Invalid, expired, or already consumed token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalidToken:
              value:
                error:
                  code: invalid_token
                  message: invalid or expired token
    UnauthorizedResponse:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            unauthorized:
              value:
                error:
                  code: unauthorized
                  message: unauthorized
...
    CSRFInvalidResponse:
      description: Missing or invalid CSRF token for cookie-authenticated mutation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            csrfInvalid:
              value:
                error:
                  code: csrf_invalid
                  message: invalid csrf token
...
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            notFound:
              value:
                error:
                  code: not_found
                  message: resource not found
...
    InternalErrorResponse:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                error:
                  code: internal_error
                  message: internal server error
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: invalid_input
        message:
          type: string
          example: invalid input
        details:
          nullable: true
          description: Optional structured error details.
...
    VerifyEmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    VerifyEmailConfirmRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 20
          example: verify_3e85ce8092e64c79a7e0b0b15da20148
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    PasswordResetConfirmRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          minLength: 20
          example: reset_89e2ec8252f24b9fa8f0fd22ad4e2f0d
        new_password:
          type: string
          minLength: 8
          maxLength: 72
          example: NewSecure123
...
    User:
      type: object
      required:
        [id, email, role, is_active, is_email_verified, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 7d4c4e98-42ba-4f55-86db-9f3d2772e6c1
        email:
          type: string
          format: email
          example: user@example.com
        full_name:
          type: string
          example: Ivan Petrov
        role:
          type: string
          enum: [customer, admin]
          example: customer
        is_active:
          type: boolean
          example: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-03-13T18:15:00Z'
        is_email_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-03-13T18:10:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-03-13T18:15:00Z'
...
    AcceptedEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [accepted]
          properties:
            accepted:
              type: boolean
              example: true
    RevokedEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [revoked]
          properties:
            revoked:
              type: boolean
              example: true
    RevokedAllEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [revoked_all]
          properties:
            revoked_all:
              type: boolean
              example: true
    VerifyEmailEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [verified, user]
          properties:
            verified:
              type: boolean
              example: true
            user:
              $ref: '#/components/schemas/User'
    PasswordResetEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [password_reset]
          properties:
            password_reset:
              type: boolean
              example: true
...